Добавлен UsersService исправлен переход в экран редактирования товара
