+++ Frontend/src/app/app.module.ts
@@ -4,15 +4,17 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
 import { AddFormComponent } from './add-form/add-form.component';
 import { ProductsComponent } from './products/products.component';
-import {FormsModule} from "@angular/forms";
+import { FormsModule } from "@angular/forms";
 import { CabinetComponent } from './cabinet/cabinet.component';
 import { HomeComponent } from './home/home.component';
-import {ProductsService} from "./products.service";
-import {AppRoutingModule} from "./app-routing.module";
+import { ProductsService } from "./products.service";
+import { UsersService } from "./users.service";
+import { AppRoutingModule } from "./app-routing.module";
 import { UserComponent } from './user/user.component';
 import { BasketComponent } from './basket/basket.component';
 import { ProductComponent } from './product/product.component';
 import { AddProductComponent } from './add-product/add-product.component';
+import { HttpClientModule } from "@angular/common/http";
 
 @NgModule({
   declarations: [
@@ -29,7 +31,8 @@ import { AddProductComponent } from './add-product/add-product.component';
   imports: [
     BrowserModule,
     FormsModule,
-    AppRoutingModule
+    AppRoutingModule,
+    HttpClientModule
   ],
   providers: [],
   bootstrap: [AppComponent]
