+++ Frontend/src/app/basket/basket.component.html
@@ -7,7 +7,7 @@
         <h2 class="product-title"><a [routerLink]="['/products', p.id]">{{p.name}}</a></h2>
         <p class="product-description">{{p.description}}</p>
         <div class="purchase-elements">
-          <p class="product-price">{{p.price}} Р</p>
+          <p class="product-price">{{p.cost}} Р</p>
         </div>
       </div>
     </div>
