+++ Frontend/src/app/product/product.component.html
@@ -24,7 +24,7 @@
     type="text"
     class="form-control"
     placeholder="Цена"
-    [(ngModel)]="price"
+    [(ngModel)]="cost"
   >
 
 </div>
