+++ Frontend/src/app/products/products.component.scss
@@ -50,6 +50,9 @@
 .btn-buy {
   margin-left: 50px;
 }
+.btn-edit {
+  margin-right: 50px;
+}
 .btn-make-purchase {
   display: inline-block;
   margin-top: 20px;
