+++ Frontend/src/app/products/products.component.ts
@@ -1,6 +1,7 @@
 import {Component, ContentChild, ElementRef, Input, OnInit} from '@angular/core';
 import {Router} from "@angular/router";
 import {Product, ProductsService} from "../products.service";
+import {UsersService} from "../users.service";
 
 @Component({
   selector: 'app-products',
@@ -11,7 +12,7 @@ export class ProductsComponent implements OnInit {
 
   @Input() product!: Product;
 
-  constructor(public productsService: ProductsService, private router: Router) { }
+  constructor(public usersService: UsersService, public productsService: ProductsService, private router: Router) { }
 
   ngOnInit() {
   }
